Fix Avalonia DataGrid wiring, Grid spacing, and OddsConverter test
- Avalonia 11 ships DataGrid as a separate package; add a reference to
  Avalonia.Controls.DataGrid 11.2.1 and include its Fluent theme in
  App.axaml. This resolves all the AVLN2000 "DataGridTextColumn / Header
  / Binding / IsReadOnly on Unknown type" errors at once.
- Avalonia's Grid does not expose RowSpacing / ColumnSpacing (those are
  WPF/MAUI-only). Strip them from every view; layouts still render
  correctly relying on RowDefinitions and child margins.
- Avalonia's DataGrid has no CornerRadius property either. Remove from
  the six DataGrid declarations; the Border CornerRadius="6" instances
  on result cards remain unchanged.
- DecimalToAmerican_round_trips: decimal literals aren't permitted in
  attribute arguments (CS0182). Pass the values as doubles and cast
  inside the test.

+++ tests/PlusEV.Tests/OddsConverterTests.cs
@@ -18,13 +18,15 @@ public class OddsConverterTests
     }
 
     [Theory]
-    [InlineData(1.91m, -110)]
-    [InlineData(2.00m, 100)]
-    [InlineData(2.50m, 150)]
-    [InlineData(1.50m, -200)]
-    public void DecimalToAmerican_round_trips(decimal dec, int expected)
+    [InlineData(1.91, -110)]
+    [InlineData(2.00, 100)]
+    [InlineData(2.50, 150)]
+    [InlineData(1.50, -200)]
+    public void DecimalToAmerican_round_trips(double dec, int expected)
     {
-        OddsConverter.DecimalToAmerican(dec).Should().Be(expected);
+        // decimal literals aren't allowed in attribute arguments (CS0182), so the
+        // theory passes doubles and we round-trip through decimal here.
+        OddsConverter.DecimalToAmerican((decimal)dec).Should().Be(expected);
     }
 
     [Fact]
